requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
alembic==1.8.0
anyio==3.6.1; python_full_version >= '3.6.2'
asgiref==3.5.2; python_version >= '3.7'
backports.cached-property==1.0.2; python_full_version >= '3.6.0'
click==8.1.3; python_version >= '3.7'
fastapi==0.78.0
graphql-core==3.2.1; python_version < '4' and python_full_version >= '3.6.0'
greenlet==1.1.2; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.13.0; python_full_version >= '3.6.0'
idna==3.3; python_version >= '3.5'
mako==1.2.0; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
psycopg2-binary==2.9.3
pydantic==1.9.1
pygments==2.12.0; python_full_version >= '3.6.0'
pyjwt==1.7.1
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-decouple==3.3
python-multipart==0.0.5
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
sqlalchemy==1.4.37
starlette==0.19.1
strawberry-graphql[debug-server]==0.114.2
typing-extensions==4.2.0; python_version < '3.10'
uvicorn==0.17.6
